Paragraph text can now explicitly be spanned
Background:
Consumers of the library may for instance want line spacing for the
paragraph text but TextView's `lineSpacingMultiplier` and/or
`lineSpacingExtra` apply line spacing to all lines.

With this change, paragraphs may explicitly be spanned with a paragraph
affecting span such as `LineHeightSpan` while the other tags
(e.g. headings) will be unaffected.